update customer service

+++ PBL3/Models/Item.cs
@@ -11,7 +11,7 @@ namespace PBL3.Models
         public required string itemName { get; set; } = string.Empty;
         public required string itemType { get; set; } = string.Empty;
         public Boolean isAvailable { get; set; } = true;
-        public required int price { get; set; } = 0;
+        public required double price { get; set; } = 0;
 
     }
 }
