+++ PBL3/Models/OrderDetails.cs
@@ -10,7 +10,7 @@ namespace PBL3.Models
         public required int itemID { get; set; } = 0;
         public required string size { get; set; } = string.Empty;
         public required int quantity { get; set; } = 0;
-        public required int priceAtOrder { get; set; } = 0;
+        public required double priceAtOrder { get; set; } = 0;
         public required string note { get; set; } = string.Empty;
     }
 }
